Rename tables_demo_api.py file
